// pslpwcl/src/app/modules/pslpapi/model/profilingQuant.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApplicazioneMin } from './applicazioneMin';
import { ClasseQuantMin } from './classeQuantMin';
import { CondizOccupazionale } from './condizOccupazionale';
import { Cpi } from './cpi';
import { DatoreLavMin } from './datoreLavMin';
import { Fontedato } from './fontedato';
import { LavAnagrafica } from './lavAnagrafica';
import { LivelloQualifMin } from './livelloQualifMin';
import { PresenzaInItaliaMin } from './presenzaInItaliaMin';
import { ProfilingQuantMin } from './profilingQuantMin';
import { Provincia } from './provincia';
import { SettoreMin } from './settoreMin';
import { Studio } from './studio';
import { TipoRichiestaMin } from './tipoRichiestaMin';
import { TitStudioIsfolMin } from './titStudioIsfolMin';

export interface ProfilingQuant { 
    codEnteIntermediario?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    codiceProfiling?: string;
    dataAggiorn?: Date;
    dataCalcolo?: Date;
    dataInserim?: Date;
    descrCpi?: string;
    eta?: number;
    flgFigliACarico?: string;
    flgGenere?: string;
    flgLavoratoAnnoPrec?: string;
    flgOccupAlProf?: string;
    flgPatente?: string;
    flgTirocinio?: string;
    idSilProfilingQuant?: number;
    idSilTRegione?: string;
    numCompFamiglia?: number;
    silLavAnagrafica?: LavAnagrafica;
    silProfilingQuantMin?: ProfilingQuantMin;
    silTApplicazioneMin?: ApplicazioneMin;
    silTCondizOccupazionale?: CondizOccupazionale;
    silTCpiOperatore?: Cpi;
    silTDatoreLavMin?: DatoreLavMin;
    silTFontedato?: Fontedato;
    silTLivelloQualifMin?: LivelloQualifMin;
    silTPresenzaInItaliaMin?: PresenzaInItaliaMin;
    silTProfilingQuantMin?: ClasseQuantMin;
    silTProvincia?: Provincia;
    silTSettoreMin?: SettoreMin;
    silTStudio?: Studio;
    silTTipoRichiestaMin?: TipoRichiestaMin;
    silTTitStudioIsfolMin?: TitStudioIsfolMin;
}