// pslpwcl/src/app/modules/pslpapi/model/studio.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { GradoStudio } from './gradoStudio';

export interface Studio { 
    codMinistero?: string;
    dataFine?: Date;
    dataInizio?: Date;
    descr?: string;
    dsDurata?: string;
    dsSinonimoTitoloStudio?: string;
    dsTipoScuola?: string;
    flgSelezInput?: string;
    id?: string;
    idSilTStudioDefCancellare?: string;
    idSilTStudioOld?: string;
    numLivello?: number;
    silTGradoStudio?: GradoStudio;
}