// pslpwcl/src/app/modules/pslpapi/model/formInserisciDid.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CallInfo } from './callInfo';
import { CondizOccupazionale } from './condizOccupazionale';
import { LavSapDid } from './lavSapDid';
import { MotivoStatoDid } from './motivoStatoDid';
import { PoliticaA02 } from './politicaA02';
import { PresenzaInItaliaMin } from './presenzaInItaliaMin';
import { ProfilingQuant } from './profilingQuant';
import { StatoOccupazionale } from './statoOccupazionale';
import { Studio } from './studio';

export interface FormInserisciDid { 
    abilitaDataPresuntoLicenziamento?: boolean;
    azioneGestioneStatoDid?: string;
    callInfo?: CallInfo;
    codFisc?: string;
    cognome?: string;
    condizOccupazionale?: CondizOccupazionale;
    dataAnzianita?: Date;
    dataDid?: Date;
    dataDidForzataDaDidINPS?: boolean;
    dataInizioSospensione?: Date;
    dataPresuntoLicenziamento?: Date;
    dataStatoDid?: Date;
    dsNote?: string;
    durataPresenzaInItalia?: PresenzaInItaliaMin;
    entePromotore?: number;
    idDidINPS?: number;
    idSilLavAnagrafica?: number;
    idSilLavSapDid?: number;
    idStatoSapDidSelezionato?: number;
    iscrittoCOMI?: boolean;
    lavSapDid?: LavSapDid;
    lavTitoloStudio?: Studio;
    listPoliticaA02?: Array<PoliticaA02>;
    mesiAnzianita?: number;
    motivoStatoDid?: MotivoStatoDid;
    nome?: string;
    numTotRedditoProspettico?: number;
    numeroComponentiNucleoFamiliare?: number;
    possessoPatente?: string;
    presenzaDiFigliACarico?: string;
    profilingQuantitativo?: ProfilingQuant;
    reddito?: string;
    sovrareddito?: string;
    statoOccupazionale?: StatoOccupazionale;
    tipoSovrareddito?: string;
}